+++ packages/utils/regl-renderer/src/geometry-utils-V2/geom2ToGeometries.test.js
@@ -13,13 +13,12 @@ test('geom2ToGeometries (empty solid)', (t) => {
   const solid = {
     sides: []
   }
-  let expected = [{ color: [1, 2, 3, 4], indices: [], normals: [], positions: [], transforms: defaultTransforms, isTransparent: false }]
+  let expected = []
   let geometries = geom2ToGeometries({ color: [1, 2, 3, 4] }, solid)
   t.deepEqual(geometries, expected)
 
   // with color
   solid.color = [4, 3, 2, 1]
-  expected = [{ color: [4, 3, 2, 1], indices: [], normals: [], positions: [], transforms: defaultTransforms, isTransparent: false }]
   geometries = geom2ToGeometries({ color: [1, 2, 3, 4] }, solid)
   t.deepEqual(geometries, expected)
 
@@ -30,14 +29,6 @@ test('geom2ToGeometries (empty solid)', (t) => {
     0, 0, 3, 0,
     0, 0, 0, 2
   ]
-  expected = [{
-    color: [4, 3, 2, 1],
-    indices: [],
-    normals: [],
-    positions: [],
-    transforms: Float32Array.from(solid.transforms),
-    isTransparent: false
-  }]
   geometries = geom2ToGeometries({ color: [1, 2, 3, 4] }, solid)
   t.deepEqual(geometries, expected)
 })
@@ -52,10 +43,22 @@ test('geom2ToGeometries (solid with sides)', (t) => {
     normals: [[0, 0, -1], [0, 0, -1], [0, 0, -1], [0, 0, -1], [0, 0, -1], [0, 0, -1]],
     positions: [[0, 0, 0], [1, 0, 0], [1, 0, 0], [1, 1, 0], [1, 1, 0], [0, 0, 0]],
     transforms: defaultTransforms,
-    isTransparent: true
+    isTransparent: true,
+    type: '2d'
   }]
   const geometries = geom2ToGeometries({ color: [1, 2, 3, 0.8] }, solid)
   t.deepEqual(geometries, expected)
 })
 
-// TODO: test SUPER LARGE solid with > 65000 points
+test('geom2ToGeometries (solid with > 65000 sides)', (t) => {
+  const solid = { sides: [] }
+  for (let i = 0; i < 70000; i++) {
+    solid.sides.push([[i, i], [i + 1, i + 1]])
+  }
+  const start = solid.sides[0]
+  const end = solid.sides[solid.sides.length - 1]
+  solid.sides.push([end[1], start[0]])
+
+  const geometries = geom2ToGeometries({ color: [1, 2, 3, 0.8] }, solid)
+  t.is(geometries.length, 3)
+})