+++ packages/utils/regl-renderer/src/rendering/commands/drawGrid/multi.js
@@ -1,5 +1,9 @@
 const makeDrawMultiGrid = (regl, params) => {
-  const { size, ticks } = params
+  const defaults = {
+    size: [50, 50],
+    ticks: [10, 1]
+  }
+  const { size, ticks } = Object.assign({}, defaults, params)
   const drawMainGrid = require('./index')(regl, { size, ticks: ticks[0] })
   const drawSubGrid = require('./index')(regl, { size, ticks: ticks[1] })
   const drawGrid = (props) => {