+++ packages/web/src/most-gestures/drags.js
@@ -1,22 +1,22 @@
 const { merge } = require('most')
 
 // based on http://jsfiddle.net/mattpodwysocki/pfCqq/
-function mouseDrags (mouseDowns$, mouseUps, mouseMoves, settings) {
-  const {pixelRatio} = settings
-  return mouseDowns$.flatMap(function (md) {
+const mouseDrags = (mouseDowns$, mouseUps, mouseMoves, settings) => {
+  const { pixelRatio } = settings
+  return mouseDowns$.flatMap((md) => {
     // calculate offsets when mouse down
-    let startX = md.pageX * pixelRatio
-    let startY = md.pageY * pixelRatio
+    const startX = md.pageX * pixelRatio
+    const startY = md.pageY * pixelRatio
     // Calculate delta with mousemove until mouseup
     let prevX = startX
     let prevY = startY
 
     return mouseMoves
-      .map(function (e) {
-        let curX = e.pageX * pixelRatio
-        let curY = e.pageY * pixelRatio
+      .map((e) => {
+        const curX = e.pageX * pixelRatio
+        const curY = e.pageY * pixelRatio
 
-        let delta = {
+        const delta = {
           left: curX - startX,
           top: curY - startY,
           x: prevX - curX,
@@ -26,32 +26,32 @@ function mouseDrags (mouseDowns$, mouseUps, mouseMoves, settings) {
         prevX = curX
         prevY = curY
 
-        const normalized = {x: curX, y: curY}
-        return {originalEvents: [e], delta, normalized, type: 'mouse'}
+        const normalized = { x: curX, y: curY }
+        return { originalEvents: [e], delta, normalized, type: 'mouse' }
       })
       .takeUntil(mouseUps)
   })
 }
 
-function touchDrags (touchStarts$, touchEnds$, touchMoves$, settings) {
-  const {pixelRatio} = settings
+const touchDrags = (touchStarts$, touchEnds$, touchMoves$, settings) => {
+  const { pixelRatio } = settings
   return touchStarts$
     // 2020-09 FIX
-    .filter(t => (t.touches.length === 1 || t.touches.length === 3)) // length 2 is pinch (zoom)
+    .filter((t) => (t.touches.length === 1 || t.touches.length === 3)) // length 2 is pinch (zoom)
     // 2020-09 FIX
-    .flatMap(function (e) {
-      let startX = e.touches[0].pageX * pixelRatio
-      let startY = e.touches[0].pageY * pixelRatio
+    .flatMap((e) => {
+      const startX = e.touches[0].pageX * pixelRatio
+      const startY = e.touches[0].pageY * pixelRatio
 
       let prevX = startX
       let prevY = startY
 
       return touchMoves$
-        .map(function (e) {
-          let curX = e.touches[0].pageX * pixelRatio
-          let curY = e.touches[0].pageY * pixelRatio
+        .map((e) => {
+          const curX = e.touches[0].pageX * pixelRatio
+          const curY = e.touches[0].pageY * pixelRatio
 
-          let delta = {
+          const delta = {
             left: curX - startX,
             top: curY - startY,
             x: prevX - curX,
@@ -61,8 +61,8 @@ function touchDrags (touchStarts$, touchEnds$, touchMoves$, settings) {
           prevX = curX
           prevY = curY
 
-          const normalized = {x: curX, y: curY}
-          return {originalEvents: [e], delta, normalized, type: 'touch'}
+          const normalized = { x: curX, y: curY }
+          return { originalEvents: [e], delta, normalized, type: 'touch' }
         })
         .takeUntil(touchEnds$)
     })
@@ -70,7 +70,7 @@ function touchDrags (touchStarts$, touchEnds$, touchMoves$, settings) {
 
 /* drag move interactions press & move(continuously firing)
 */
-function drags ({mouseDowns$, mouseUps$, mouseMoves$, touchStarts$, touchEnds$, longTaps$, touchMoves$}, settings) {
+const drags = ({ mouseDowns$, mouseUps$, mouseMoves$, touchStarts$, touchEnds$, longTaps$, touchMoves$ }, settings) => {
   // 2020-09 FIX
   // touchMoves$ = touchMoves$.filter(t => t.touches.length === 1)
   // 2020-09 FIX
@@ -86,4 +86,4 @@ function drags ({mouseDowns$, mouseUps$, mouseMoves$, touchStarts$, touchEnds$, 
   return drags$
 }
 
-module.exports = {mouseDrags, touchDrags, drags}
+module.exports = { mouseDrags, touchDrags, drags }