+++ packages/web/src/sideEffects/i18n/index.js
@@ -1,10 +1,13 @@
-const callBackToStream = require('../../observable-utils/callbackToObservable')
 const isElectron = require('is-electron')
+
+const callBackToStream = require('../../most-utils/callbackToObservable')
+
 const longNames = {
   en: 'english',
   de: 'german',
   fr: 'french',
-  ja: 'japanese'
+  ja: 'japanese',
+  hr: 'croatian'
 }
 
 const getDefaultLocale = () => {